Don't delete hosts we have a contract with from hostdb

+++ modules/renter/hostdb/scan.go
@@ -275,8 +275,10 @@ func (hdb *HostDB) updateEntry(entry modules.HostDBEntry, netErr error) {
 
 	// If the host has been offline for too long, delete the host from the
 	// hostdb. Only delete if there have been enough scans over a long enough
-	// period to be confident that the host really is offline for good.
-	if time.Now().Sub(newEntry.ScanHistory[0].Timestamp) > maxHostDowntime && !recentUptime && len(newEntry.ScanHistory) >= minScans {
+	// period to be confident that the host really is offline for good, and if we
+	// don't have any contracts with that host.
+	_, haveContractWithHost := hdb.knownContracts[entry.PublicKey.String()]
+	if !haveContractWithHost && time.Now().Sub(newEntry.ScanHistory[0].Timestamp) > maxHostDowntime && !recentUptime && len(newEntry.ScanHistory) >= minScans {
 		// Remove from hosttrees
 		err := hdb.remove(newEntry.PublicKey)
 		if err != nil {
